% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeViz.R
\name{viz_tree}
\alias{viz_tree}
\title{Tree Vizualization}
\usage{
viz_tree(data = data.frame, tree = rpart())
}
\arguments{
\item{data}{a data frame whose first 3 columns are in the form:
Outcome, Feature 1, Feature 2}

\item{tree}{a classification or regression tree made with the rpart package}
}
\value{
a ggplot object
}
\description{
Draw lines on a scatter plot to visualize tree decisions
}
\examples{
# This example uses the mnist dataset from the dslabs packages
library(dslabs)
data(mnist_27)
train27<-mnist_27$train
classTree<- rpart(y~x_1+x_2, data = train27, method = "class")
viz_tree(train27, classTree)
}
